+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [3.5.0] - Unreleased
+
+### Added
+- Custom Endpoints
+- New `DynamicSelect` input component for custom fields
+- `customEndpointsMenuItemText` property for language dictionary
+
 ## [3.4.0] - 2018-10-25
 
 ### Added